# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Product Scraper API - scrapes e-commerce product data and exports to CSV
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ScrapeJobInput {
  /**
     * Number of pages to scrape (each page has ~20 books)
     * @minimum 1
     * @maximum 50
     */
  pages: number;
  /**
     * Optional category slug to filter scraping
     * @nullable
     */
  category?: string | null;
}

export type ScrapeJobStatus = typeof ScrapeJobStatus[keyof typeof ScrapeJobStatus];


export const ScrapeJobStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface ScrapeJob {
  id: number;
  status: ScrapeJobStatus;
  pagesRequested: number;
  pagesScraped?: number;
  productsFound?: number;
  startedAt: string;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  error?: string | null;
}

export interface Product {
  id: number;
  title: string;
  price: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  availability: string;
  category: string;
  url: string;
  scrapedAt: string;
  /** @nullable */
  jobId?: number | null;
}

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface RatingCount {
  rating: number;
  count: number;
}

export interface ProductStats {
  totalProducts: number;
  avgPrice: number;
  avgRating: number;
  totalCategories: number;
  byRating: RatingCount[];
  /** @nullable */
  lastScrapedAt: string | null;
}

export interface CategoryCount {
  category: string;
  count: number;
}

export interface ClearResult {
  deleted: number;
}

export type ListProductsParams = {
page?: number;
limit?: number;
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
minRating?: number | null;
/**
 * @nullable
 */
maxPrice?: number | null;
sort?: ListProductsSort;
/**
 * @nullable
 */
search?: string | null;
};

export type ListProductsSort = typeof ListProductsSort[keyof typeof ListProductsSort];


export const ListProductsSort = {
  price_asc: 'price_asc',
  price_desc: 'price_desc',
  rating_desc: 'rating_desc',
  title_asc: 'title_asc',
} as const;

export type ExportProductsCsvParams = {
/**
 * @nullable
 */
category?: string | null;
/**
 * @nullable
 */
minRating?: number | null;
};