# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Product Scraper API - scrapes e-commerce product data and exports to CSV
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  title: string;
  price: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  rating: number;
  availability: string;
  category: string;
  url: string;
  scrapedAt: Date;
  /** @nullable */
  jobId?: number | null;
}