# lib/api-zod/src/generated/types/productListResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Product Scraper API - scrapes e-commerce product data and exports to CSV
 * OpenAPI spec version: 0.1.0
 */
import type { Product } from './product';

export interface ProductListResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}