# lib/api-zod/src/generated/types/productStats.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Product Scraper API - scrapes e-commerce product data and exports to CSV
 * OpenAPI spec version: 0.1.0
 */
import type { RatingCount } from './ratingCount';

export interface ProductStats {
  totalProducts: number;
  avgPrice: number;
  avgRating: number;
  totalCategories: number;
  byRating: RatingCount[];
  /** @nullable */
  lastScrapedAt: Date | null;
}