# lib/api-zod/src/generated/types/scrapeJob.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Product Scraper API - scrapes e-commerce product data and exports to CSV
 * OpenAPI spec version: 0.1.0
 */
import type { ScrapeJobStatus } from './scrapeJobStatus';

export interface ScrapeJob {
  id: number;
  status: ScrapeJobStatus;
  pagesRequested: number;
  pagesScraped?: number;
  productsFound?: number;
  startedAt: Date;
  /** @nullable */
  completedAt?: Date | null;
  /** @nullable */
  error?: string | null;
}