# lib/api-zod/src/generated/types/scrapeJobInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Product Scraper API - scrapes e-commerce product data and exports to CSV
 * OpenAPI spec version: 0.1.0
 */

export interface ScrapeJobInput {
  /**
     * Number of pages to scrape (each page has ~20 books)
     * @minimum 1
     * @maximum 50
     */
  pages: number;
  /**
     * Optional category slug to filter scraping
     * @nullable
     */
  category?: string | null;
}